requirements.txt
selenium>=4.15.0
pandas>=2.0.0
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
python-dotenv>=1.0.0


